+++ packages/sources/src/components/map-controls/map-controls.content.ts
@@ -1,5 +1,28 @@
-export const baseLayers = [{ id: 0, name: 'Kaart' }, { id: 1, name: 'Luchtfoto' }];
+export const baseLayers = [
+  {
+    id: 0,
+    name: 'Kaart'
+  },
+  {
+    id: 1,
+    name: 'Luchtfoto',
+    checked: true
+  }
+];
 
-export const selectedBaseLayer = baseLayers[baseLayers.length - 1].id;
-
-export const overlays = [{ id: 0, name: 'Kadastrale grenzen' }, { id: 1, name: 'Basisregistratie Adressen en Gebouwen (BAG)' }];
+export const overlays = [
+  {
+    id: 0,
+    name: 'Kadastrale grenzen'
+  },
+  {
+    id: 1,
+    name: 'Basisregistratie Adressen en Gebouwen (BAG)',
+    checked: true
+  },
+  {
+    id: 2,
+    name: 'Riool',
+    disabled: true
+  }
+];